Rename grid.js example to mnist, and do small cleanup
