+++ examples/mnist/README.md
@@ -1,0 +1,11 @@
+# Syft.js MNIST Example
+
+Prerequisites:
+
+1. Execute `npm install`
+2. Install and start [PyGrid](https://github.com/OpenMined/pygrid)
+3. Install [PySyft](https://github.com/OpenMined/PySyft) and execute "Create Plan" and "Host Plan" notebooks from `examples/experimental/FL Training Plan` folder.
+
+Finally, execute `npm start`.
+
+This will launch a web browser running the file `index.js`. Every time you make changes to this file, the server will automatically re-compile your code and refresh the page. No need to start and stop. :)
