+++ examples/mnist/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "with-grid",
+  "name": "syftjs-mnist",
   "version": "1.0.0",
   "private": true,
   "description": "",
