+++ examples/mnist/webpack.config.js
