--- examples/with-grid/README.md
@@ -1,8 +1,0 @@
-# syft.js/grid.js Example
-
-In order to get this running, you simply need to run the following:
-
-1. `npm install`
-2. `npm start`
-
-This will launch a web browser running the file `index.js`. Every time you make changes to this file, the server will automatically re-compile your code and refresh the page. No need to start and stop. :)
